requirements/dev-requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements/dev-requirements.in
#
blinker==1.9.0
    # via flask
click==8.2.1
    # via flask
factory-boy==3.3.3
    # via
    #   -r requirements/dev-requirements.in
    #   pytest-factoryboy
faker==37.4.0
    # via factory-boy
flask==3.1.1
    # via -r /home/robert/apps/photo_crop_app/requirements/requirements.in
inflection==0.5.1
    # via pytest-factoryboy
iniconfig==2.1.0
    # via pytest
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   werkzeug
packaging==25.0
    # via
    #   pytest
    #   pytest-factoryboy
pluggy==1.6.0
    # via pytest
pygments==2.19.2
    # via pytest
pytest==8.4.1
    # via
    #   -r requirements/dev-requirements.in
    #   pytest-factoryboy
    #   pytest-mock
pytest-factoryboy==2.8.1
    # via -r requirements/dev-requirements.in
pytest-mock==3.14.1
    # via -r requirements/dev-requirements.in
ruff==0.12.1
    # via -r requirements/dev-requirements.in
typing-extensions==4.14.0
    # via pytest-factoryboy
tzdata==2025.2
    # via faker
werkzeug==3.1.3
    # via flask
